notesApp(DjangoRest, React)/requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.2-